refactor: Rename macro from generate_function_declaration to tool and modularize
- Rename #[generate_function_declaration] to #[tool] for brevity
- Split macro implementation into separate modules: parsing, schema, codegen
- Add trybuild for compile-time UI tests
- Update all examples and tests to use new #[tool] attribute

Code quality improvements from PR review:
- Replace expect() with proper compile errors for schema serialization
- Handle empty properties case explicitly instead of silent fallback
- Fix u64 number handling without runtime unwrap()
- Add validation for unknown attribute keys (typo detection)
- Add validation for non-existent parameter references
- Fix clippy lint: use let-chains for collapsible if statements
- Add module-level and function-level documentation
- Update CLAUDE.md and README.md with new macro name
- Add UI test for non-existent parameter validation
