+++ tests/ui/fail_invalid_attr.rs
@@ -1,0 +1,10 @@
+use rust_genai_macros::tool;
+
+// This should fail because "invalid_key" is not a recognized attribute.
+// Valid attributes are: description, enum_values
+#[tool(name(invalid_key = "value"))]
+fn test_invalid_attr(name: String) -> String {
+    name
+}
+
+fn main() {}
